+++ CHANGELOG.md
@@ -36,6 +36,10 @@
 
 ### Bug Fixes
 
+* Fix `vertical_whitespace_between_cases` false positives when comments separate cases.  
+  [leno23](https://github.com/leno23)
+  [#6412](https://github.com/realm/SwiftLint/issues/6412)
+
 * Avoid false positives in `prefer_self_in_static_references` for generic
   constraints and generic parameter bounds such as `where A: P` and `<A: P>`
   in classes and extensions.